seems to be working fine but weird os error is causing tests to fail
